Move Labels implementation to its own file
